HBASE-5945 Reduce buffer copies in IPC server response path
git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.96@1543463 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-hadoop-compat/src/main/java/org/apache/hadoop/hbase/ipc/MetricsHBaseServerSource.java
@@ -62,7 +62,7 @@ public interface MetricsHBaseServerSource extends BaseSource {
 
   void authenticationFailure();
 
-  void sentBytes(int count);
+  void sentBytes(long count);
 
   void receivedBytes(int count);
