+++ hbase-hadoop1-compat/src/main/java/org/apache/hadoop/hbase/ipc/MetricsHBaseServerSourceImpl.java
@@ -86,7 +86,7 @@ public class MetricsHBaseServerSourceImpl extends BaseSourceImpl
   }
 
   @Override
-  public void sentBytes(int count) {
+  public void sentBytes(long count) {
     this.sentBytes.incr(count);
   }
